requirements.txt
discord
selenium
webdriver-manager
python-dotenv